Refractor Products

+++ src/main/kotlin/game/product/BasicProductPile.kt
@@ -1,7 +1,7 @@
 package game.product
 
 data class BasicProductPile(
-    override val maximum: Long,
+    override val maximum: Int,
     override var productCount: Int,
     override val type: ProductType,
     override var price: Int
